Fix crash on pathological case
There are many ways in which kallisto might not find any reads that
pseudo-align to a candidate transcripts from a subgraph. Some of these
failure paths are already accounted for, but this is a new one. Absent
enough understanding to detect this category of pathological cases before
running kallisto, I have chosen to instead detect when kallisto fails
to pseudo-align any reads and silently drop subgraphs affected by this
error from further processing.